requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
mysql-connector-python==8.2.0
pandas==2.1.4
numpy==1.26.3
pytest==7.4.3
httpx==0.26.0